chore: updated example app
